requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

langchain==0.2.14
langgraph==0.2.28

openai==1.40.0
httpx==0.27.0
requests==2.32.3
tiktoken==0.7.0
pydantic>=2.7,<3
groq==0.11.0